requirements.txt
accelerate==0.32.1
aiofiles==23.2.1
aiogram==3.18.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chatgpt-wrapper==0.0.1
click==8.2.1
colorama==0.4.6
comtypes==1.4.10
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
datasets==2.20.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docx2pdf==0.1.8
fake-useragent==2.2.0
filelock==3.15.4
Flask==3.1.1
fonttools==4.55.0
frozenlist==1.4.1
fsspec==2024.5.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.24.0
idna==3.10
instaloader==4.12.1
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
keyboard==0.13.5
kiwisolver==1.4.7
lxml==5.2.2
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mkl==2021.4.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mutagen==1.47.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
openai==0.28.0
OpenAIAuth==3.0.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.2
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.0.0
playwright==1.52.0
prompt_toolkit==3.0.50
psutil==7.0.0
pyaes==1.6.1
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
PyMuPDF==1.25.5
pyparsing==3.2.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
PyRect==0.2.0
Pyrogram==2.0.106
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-telegram-bot==22.1
pytube==15.0.0
pytweening==1.2.0
pytz==2024.1
pywin32==308
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
revChatGPT==6.8.6
rich==13.9.4
rsa==4.9
safetensors==0.4.3
selenium==4.11.2
setuptools==75.8.0
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.6
sympy==1.13.1
tbb==2021.13.0
Telethon==1.38.1
termcolor==2.5.0
tiktoken==0.9.0
tls-client==1.0.1
tokenizers==0.19.1
torch==2.3.1+cpu
tqdm==4.66.4
transformers==4.42.4
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==2.4.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.1.3
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
youtube-dl==2021.12.17
yt-dlp==2024.8.6
